+++ src/api/Auth/DeleteUser.js
@@ -1,4 +1,4 @@
-import { Axios } from './Axios';
+import { Axios } from '../Axios';
 
 export const DeleteUser = (callbackFunction) => {
   Axios.delete(`/users`, {})
