+++ src/api/Auth/LogoutApi.js
@@ -1,4 +1,4 @@
-import { Axios } from './Axios';
+import { Axios } from '../Axios';
 
 export const LogoutApi = (data, callbackFunctions) => {
   const { navigateSuccess, navigateError } = callbackFunctions;
