+++ src/api/Game/GetChapters.js
@@ -1,7 +1,7 @@
-import { Axios } from './Axios';
+import { Axios } from '../Axios';
 
-export const GetRate = (type, callbackFunction) => {
-  Axios.get(`/learnings/progress`, {
+export const GetChapters = (type, callbackFunction) => {
+  Axios.get(`/learnings/chapters`, {
     params: {
       type: type,
     },
