+++ src/api/Game/PostScoring.js
@@ -1,4 +1,4 @@
-import { Axios } from './Axios';
+import { Axios } from '../Axios';
 
 export const PostScoring = (
   problemId,
