--- src/assets/data/Content_Choice_Dummy.json
@@ -1,29 +1,0 @@
-[
-  {
-    "id": "a",
-    "type": "a",
-    "title": "정의 검색",
-    "scenario": "멋사에 합격했다. 공부를 시작해보자! Git 이라는 것을 설치 해오는것이 과제이다. 그런데 Git이 뭐지?",
-    "question": " → 검색 해야하는 키워드는 무엇일까?",
-    "help": "도움말",
-    "option": ["Git", "공부", "멋사"]
-  },
-  {
-    "id": "b",
-    "type": "a",
-    "title": "정의 검색",
-    "scenario": "사진에 대한 열정이 생겨 DSLR 카메라를 구매하게 됐다. 첫 번째로 해야 할 일은 셔터 스피드를 조정하는 것이라는데, 셔터 스피드가 뭐지?",
-    "question": " → 검색 해야하는 키워드는 무엇일까?",
-    "help": "도움말",
-    "option": ["사진", "셔터 스피더", "DSLR"]
-  },
-  {
-    "id": "c",
-    "type": "a",
-    "title": "정의 검색",
-    "scenario": "피트니스에 관심이 생겨서 개인 트레이닝을 시작하기로 결정했다. 트레이너가 스쿼트를 해보라고 하는데, 스쿼트가 뭐지?",
-    "question": " → 검색 해야하는 키워드는 무엇일까?",
-    "help": "도움말",
-    "option": ["피트니스", "트레이닝", "스쿼트"]
-  }
-]
