--- src/assets/data/Content_Short_Dummy.json
@@ -1,9 +1,0 @@
-[
-  {
-    "type": "b",
-    "title": "키워드 도출",
-    "scenario": "오늘은 새로운 요리를 해보려고 합니다. 떡볶이를 만들고 싶은데 재료들이 많이 필요할 것 같고 어떤 과정을 거쳐야 하는지도 잘 모르겠습니다. 레시피를 얻고 싶어요.",
-    "question": " → 어떤 검색어를 입력해야 할까요?",
-    "help": "도움말"
-  }
-]
