--- src/assets/data/Rate_DummyData_Basic.json
@@ -1,5 +1,0 @@
-{
-  "statusCode": 200,
-  "message": "Progress successfully retrieved",
-  "progress": 30
-}
