--- src/assets/data/SideBar_DummyData_Advanced.json
@@ -1,48 +1,0 @@
-{
-  "statusCode": 200,
-  "message": "Chapters successfully retrieved",
-  "chapters": [
-    {
-      "id": "4d34615a-50be-4dac-9a76-1558d749bde2",
-      "title": "AND, OR, NOT 사용",
-      "help_message": "blah",
-      "isCompleted": true
-    },
-    {
-      "id": "b18b3b5b-4dbc-4a9a-89d6-4e61aa4ceb22",
-      "title": "사이트 특정 검색",
-      "help_message": null,
-      "isCompleted": true
-    },
-    {
-      "id": 3,
-      "title": "부정 연산자 사용 (-)",
-      "help_message": "blah",
-      "isCompleted": true
-    },
-    {
-      "id": 4,
-      "title": "파일 유형 검색 사용",
-      "help_message": "blah",
-      "isCompleted": true
-    },
-    {
-      "id": 5,
-      "title": "와일드 카드 사용 (*)",
-      "help_message": "blah",
-      "isCompleted": true
-    },
-    {
-      "id": 6,
-      "title": "제목1",
-      "help_message": "blah",
-      "isCompleted": false
-    },
-    {
-      "id": 7,
-      "title": "제목2",
-      "help_message": "blah",
-      "isCompleted": false
-    }
-  ]
-}
