--- src/assets/data/SideBar_DummyData_Basic.json
@@ -1,60 +1,0 @@
-{
-  "statusCode": 200,
-  "message": "Chapters successfully retrieved",
-  "chapters": [
-    {
-      "id": "4d34615a-50be-4dac-9a76-1558d749bde8",
-      "title": "챕터 1",
-      "help_message": "blah",
-      "isCompleted": true
-    },
-    {
-      "id": "b18b3b5b-4dbc-4a9a-89d6-4e61aa4ceb2c",
-      "title": "챕터 2",
-      "help_message": null,
-      "isCompleted": true
-    },
-    {
-      "id": 3,
-      "title": "키워드 조합",
-      "help_message": "blah",
-      "isCompleted": true
-    },
-    {
-      "id": 4,
-      "title": "날씨",
-      "help_message": "blah",
-      "isCompleted": true
-    },
-    {
-      "id": 5,
-      "title": "사전",
-      "help_message": "blah",
-      "isCompleted": true
-    },
-    {
-      "id": 6,
-      "title": "계산",
-      "help_message": "blah",
-      "isCompleted": false
-    },
-    {
-      "id": 7,
-      "title": "단위변환",
-      "help_message": "blah",
-      "isCompleted": false
-    },
-    {
-      "id": 8,
-      "title": "스포츠",
-      "help_message": "blah",
-      "isCompleted": false
-    },
-    {
-      "id": 9,
-      "title": "요약 정보",
-      "help_message": "blah",
-      "isCompleted": false
-    }
-  ]
-}
