+++ src/components/Game/Type_FillBlank.js
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
-import search_icon from '../../assets/images/search_icon.svg';
-import SquareButton from '../../components/SquareButton';
 import { useEffect, useState } from 'react';
+import SquareButton from './SquareButton';
 import next_icon from '../../assets/images/right_arrow_icon.svg';
+import search_icon from '../../assets/images/search_icon.svg';
 
 const TypeFillBlank = ({
   defaultMsg,
@@ -63,7 +63,7 @@ const TypeFillBlank = ({
 
 const FillBlankWrapper = styled.div`
   width: 100%;
-  height: 100%;
+  height: 250px;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
