+++ src/components/Main/MainBanner.js
