+++ src/pages/Main/Main.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
-import MainBanner from './MainBanner';
+import MainBanner from '../../components/Main/MainBanner';
 import EduIcon from '../../assets/images/icon_edu.png';
 import MyPage from '../../assets/images/icon_mypage.png';
